// src/swagger/transform/models/playlist-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API SPEC FOR MOVESONG FRONTEND
 * REST API SPEC FOR MOVESONG FRONTEND
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PlaylistDto
 */
export interface PlaylistDto {
    /**
     * 
     * @type {string}
     * @memberof PlaylistDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistDto
     */
    channelTitle?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistDto
     */
    thumbnailUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof PlaylistDto
     */
    itemCount?: number;
}